Refactoring
Placed a Self get-property of all systems, and made most (if not all) methods as static to prevent from creating another property and calleding GameManager.GetSystem<GameSystem>().

Also changed any use of modifiers to the object implementing it.
Changing ((IHealthModifier)this).HealthBonus) to caching the IHealthModifier implementation to a variable (for example: IHealthModifier HealthModifier => this), and using HealthModifier.HealthBonus instead. From this point on, creating skills, items, mados, and deities should be a little more simplified.